% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNLP.R
\docType{package}
\name{MetaNLP}
\alias{MetaNLP}
\alias{MetaNLP-class}
\title{Natural Language Processing for Meta Analysis}
\usage{
MetaNLP(
  path,
  bounds = c(2, Inf),
  word_length = c(3, Inf),
  language = "english",
  ...
)
}
\arguments{
\item{path}{Path to the CSV file}

\item{bounds}{An integer vector of length 2. The first value specifies
the minimum number of appearances of a word to become a column of the word
count matrix, the second value specifies the maximum number.
Defaults to \code{c(2, Inf)}.}

\item{word_length}{An integer vector of length 2. The first value specifies
the minimum number of characters of a word to become a column of the word
count matrix, the second value specifies the maximum number.
Defaults to \code{c(3, Inf)}.}

\item{language}{The language for lemmatization and stemming. Supported
languages are \code{english}, \code{french}, \code{german}, \code{italian},
\code{portugese}, \code{romanian}, \code{russian}, \code{spanish} and
\code{swedish}.}

\item{...}{Additional arguments passed on to \code{read.csv2}.
See \link[utils]{read.table}.}
}
\value{
An object of class \code{MetaNLP}
}
\description{
The \pkg{MetaNLP} package provides methods to quickly transform a
CSV-file with titles and abstracts to an R data frame that can be
used for automatic title-abstract screening using machine learning.

A \code{MetaNLP} object is the base class of the package \pkg{MetaNLP}.
It is initialized by passing the path to a CSV file and constructs
a data frame which column names are the words that occur in the titles
and abstracts and which cells contain the word counts for each
paper.
}
\details{
An object of class \code{MetaNLP} contains a slot data_frame where
the word count data frame is stored.
The CSV file must have a column \code{ID} to identify each paper, a column
\code{title} with the belonging titles of the papers and a column
\code{abstract} which contains the abstracts. Furthermore, to store the
decision for each paper, a column \code{decision} should exist, where the
values are either "yes" and "no" or "include" and "exclude". The value "maybe"
is handled as a "yes"/"include".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imbi-heidelberg/MetaNLP}
  \item Report bugs at \url{https://github.com/imbi-heidelberg/MetaNLP/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Pilz \email{maximilian.pilz@itwm.fraunhofer.de} (\href{https://orcid.org/0000-0002-9685-1613}{ORCID})

Authors:
\itemize{
  \item Nico Bruder \email{bruder@imbi.uni-heidelberg.de}
  \item Samuel Zimmermann \email{zimmermann@imbi.uni-heidelberg.de} (\href{https://orcid.org/0009-0000-4828-9294}{ORCID})
  \item Johannes Vey \email{vey@imbi.uni-heidelberg.de} (\href{https://orcid.org/0000-0002-2610-9667}{ORCID})
}

Other contributors:
\itemize{
  \item Institute of Medical Biometry - University of Heidelberg [copyright holder]
}

}
