% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNLP.R
\name{plot,MetaNLP,ANY-method}
\alias{plot,MetaNLP,ANY-method}
\title{Create word cloud from MetaNLP-object}
\usage{
\S4method{plot}{MetaNLP,ANY}(
  x,
  max.words = 70,
  colors = c("snow4", "darkgoldenrod1", "turquoise4", "tomato"),
  ...
)
}
\arguments{
\item{x}{A MetaNLP object to plot}

\item{max.words}{Number of words in the word cloud}

\item{colors}{Character vector with the colors in}

\item{...}{Additional parameters for \link[wordcloud]{wordcloud}}
}
\description{
This method creates a word cloud from a MetaNLP object. The word size
indicates the frequency of the words.
}
\examples{
obj <- MetaNLP("test_data.csv")
plot(obj)

}
