% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNLP.R
\docType{package}
\name{MetaNLP}
\alias{MetaNLP}
\alias{MetaNLP-class}
\title{Natrual Language Processing for Meta Analysis}
\usage{
MetaNLP(path)
}
\arguments{
\item{path}{path to the CSV file}
}
\value{
an object of class \code{MetaNLP}
}
\description{
The \pkg{MetaNLP} package provides methods to quickly transform a
CSV-file with titles and abstracts to an R data frame that can be
used for automatic title-abstract screening using machine learning.

A \code{MetaNLP} object is the base class of the \pkg{MetaNLP}.
It is initialized by passing the path to a CSV file and constructs
a data frame which column names are the words that occur in the titles
and abstracts and which cells contain the word counts for each
paper.
}
\details{
An object of class \code{MetaNLP} contains a slot data_frame where
the word count data frame is stored.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imbi-heidelberg/MetaNLP}
  \item Report bugs at \url{https://github.com/imbi-heidelberg/MetaNLP/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Pilz \email{maximilian.pilz@itwm.fraunhofer.de} (\href{https://orcid.org/0000-0002-9685-1613}{ORCID})

Authors:
\itemize{
  \item Nico Bruder \email{bruder@imbi.uni-heidelberg.de}
  \item Samuel Zimmermann \email{zimmermann@imbi.uni-heidelberg.de} (\href{https://orcid.org/0009-0000-4828-9294}{ORCID})
  \item Johannes Vey \email{vey@imbi.uni-heidelberg.de} (\href{https://orcid.org/0000-0002-2610-9667}{ORCID})
}

Other contributors:
\itemize{
  \item Institute of Medical Biometry - University of Heidelberg [copyright holder]
}

}
