% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{write_csv}
\alias{write_csv}
\alias{write_csv,MetaNLP-method}
\title{Save the word count matrix}
\usage{
write_csv(object, ...)

\S4method{write_csv}{MetaNLP}(object, path = "", type = c("train", "test"), ...)
}
\arguments{
\item{object}{An object of class MetaNLP.}

\item{...}{Additional arguments for \link[utils]{write.table}.}

\item{path}{Path where to save the csv. If no path is set, the csv is saved
in the current working directory}

\item{type}{Specifies if the word count matrix should be saved as
"train_wcm.csv" or "test_wcm.csv". If the user wants to use another file name,
the whole path including the file name should be given as the \code{path}
argument}
}
\description{
This function can be used to save the word count matrix of a MetaNLP object
as a csv-file.
}
\details{
Overall, there are three options to specify the path. By
default, no path is set, so the csv is saved as "train_wcm.csv" or
"test_wcm.csv" in the current working directory. If a path to a specific
folder is given (but the path name does not end with ".csv"), the file is
saved in this folder as "train_wcm.csv" or "test_wcm.csv".
By providing a path ending with ".csv", the user can override the default
naming convention and the file is saved according to this path.
}
\examples{
\dontrun{
obj <- MetaNLP("test_data.csv")
obj2 <- delete_stop_words(obj)
write_csv(obj2)
write_csv(obj2, path = "foo.csv")
}


}
